[EC-139] Refactoring: moved Receipts rlpEncDec to PV63 and added comments
