+++ src/main/scala/io/iohk/ethereum/domain/Receipt.scala
@@ -24,18 +24,3 @@ case class Receipt(
        """.stripMargin
   }
 }
-
-object Receipt {
-  implicit val rlpEncDec = new RLPEncoder[Receipt] with RLPDecoder[Receipt] {
-    override def encode(obj: Receipt): RLPEncodeable = {
-      import obj._
-      RLPList(postTransactionStateHash, cumulativeGasUsed, logsBloomFilter, toRlpList[TxLogEntry](logs)(TxLogEntryImplicits.rlpEncDec))
-    }
-
-    override def decode(rlp: RLPEncodeable): Receipt = rlp match {
-      case RLPList(postTransactionStateHash, cumulativeGasUsed, logsBloomFilter, logs: RLPList) =>
-        Receipt(postTransactionStateHash, cumulativeGasUsed, logsBloomFilter, fromRlpList[TxLogEntry](logs)(TxLogEntryImplicits.rlpEncDec))
-      case _ => throw new RuntimeException("Cannot decode Receipt")
-    }
-  }
-}
