+++ src/main/scala/io/iohk/ethereum/network/p2p/validators/BlockValidator.scala
@@ -6,6 +6,7 @@ import io.iohk.ethereum.mpt.RLPByteArraySerializable
 import io.iohk.ethereum.network.p2p.messages.CommonMessages.SignedTransactions._
 import io.iohk.ethereum.network.p2p.messages.PV62.BlockBody
 import io.iohk.ethereum.network.p2p.messages.PV62.BlockHeaderImplicits.headerRlpEncDec
+import io.iohk.ethereum.network.p2p.messages.PV63.ReceiptImplicits.receiptRlpEncDec
 import io.iohk.ethereum.rlp.RLPImplicitConversions.toRlpList
 import io.iohk.ethereum.rlp._
 import io.iohk.ethereum.utils.ByteUtils.or
